requirements.txt
requests
pillow
google-generativeai
python-telegram-bot==13.15
urllib3==1.26.18